Inflation curves unregister in OREData.
Make the change discussed in PR #326. The change unregisters the index
from the ZC and YoY helpers and hence the inflation curve does not
observe the inflation index. It was interfering with scenario generation
as outlined in the PR. The YoY helpers and inflation curve are still not
observing the evaluationDate.